Rename sqlQueryRun.js to sql.js
