+++ sof-js/src/server.js
@@ -7,7 +7,7 @@ import { mountRoutes as mountFhirRoutes } from './server/fhir.js'
 import { mountRoutes as mountViewsRoutes } from './server/views.js'
 import { mountRoutes as mountEvaluateRoutes } from './server/evaluate.js'
 import { mountRoutes as mountValidateRoutes } from './server/validate.js'
-import { mountRoutes as mountSqlQueryRunRoutes } from './server/sqlQueryRun.js'
+import { mountRoutes as mountSqlQueryRunRoutes } from './server/sql.js'
 import { migrate, getDb } from './server/db.js'
 import { resourceTypes } from './server/utils.js'
 import { layout } from './server/ui.js'
